+++ src/utils/helpers/date/__spec__.js
@@ -89,26 +89,22 @@ describe('DateHelper', () => {
     });
   });
 
-  fdescribe('withinRange', () => {
-    let momentSpy = jasmine.createSpyObj('moment', ['format', 'add', 'subtract']);
-
-    beforeEach(() => {
-      spyOn(DateHelper, 'isValidDate').and.returnValue(true);
-      momentSpy.format.and.returnValue(momentValue);
-    });
-
+  describe('withinRange', () => {
     it('returns true if the date is within the given range', () => {
-      let testDate = moment('15/10/2015', 'DD/MM/YYYY');
-      momentSpy.add.and.returnValue(moment(momentValue).add(30, 'days').format());
-      momentSpy.subtract.and.returnValue(moment(momentValue).subtract(30, 'days').format());
+      let testDate = moment().format('DD-MM-YYYY');
       expect(DateHelper.withinRange(testDate, 30, 'days')).toBeTruthy();
     });
 
     it('returns false if the date is outside the given range', () => {
-      let testDate = moment('29/10/2015', 'DD/MM/YYYY');
-      momentSpy.add.and.returnValue(moment(momentValue).add(10, 'days').format());
-      momentSpy.subtract.and.returnValue(moment(momentValue).subtract(10, 'days').format());
+      let testDate = moment().add(100, 'days').format('DD-MM-YYYY');
       expect(DateHelper.withinRange(testDate, 10, 'days')).toBeFalsy();
+    });
+
+    describe('when date is invalid', () => {
+      it('returns true', () => {
+        let testDate = '98has9hd9h.309af/3de'
+        expect(DateHelper.withinRange(testDate, 30, 'days')).toBeTruthy();
+      });
     });
   });
 });